Job: NonDestructiveSaves - Updated configuration to utilize same document loaded from.
This job was to provide a method for reading property changes back into an existing document. This way the document saving is non-destructive, so that elements that weren't read into Simpl don't get left out, and the order of the document is preserved.
Also added code to write defaults to a file if it doesn't exist, so that a default structure is built. This way a file can be pulled from the processor ready to edit. (This still needs testing.)