Updated processing of values, allowing extra delineators to specify values within Xml tags. Various other tweaks.
Updated value handling. Placing an extra delineator character (period by default) at the end of a path parameter, the program will place (or look for) the value of the path inside the xml tags for that element.
Moved IsInitialized property to the ProcessorBase class. Added IsProcessorInitialized property that can easily be called from Simpl+.
Changed the handling of locks, moving away from a lock object and locking on the individual processor objects instead.
Removed previously added IsElement property, opting for handling the tracking of this data in individual methods. This was needed for the updated value handling with trailing delimiters.
Added some additional error checking in various places.